+++ src/components/views/Cobd.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import CodbForm from '../CodbForm.vue'
+import CodbForm from '../CoBDComponents/CodbForm.vue'
 import CodbHeader from '../CoBDComponents/CodbHeader.vue'
 import CodbFooter from '../CoBDComponents/CodbFooter.vue'
 </script>
